added input thread back to cli_controller
